zembed/requirements.txt
appdirs==1.4.0
Django==1.10.5
olefile==0.44
packaging==16.8
Pillow==4.0.0
pymongo==3.4.0
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
